Catch Reaction Issue for Private Channels
If the reaction is not in a guild or the member is not cached, `reactor` will be an object with an id key; no other property is guaranteed. Currently this plugin breaks while working on my plugin-plugin which has message reactions in private channels, and this addition would solve the issue!